fix(settings): stop event propagation from tappable elements
This fixes an issue introduced in commit 0d35a98.

Event bubbling from the reset button to the setting itself was
initially blocked inside the `click` listener of the button. However,
this stopped working after using the `trigger` custom event instead.

Event bubbling is now blocked on all tappable elements, both as a fix
and as a preventive measure to avoid possible future issues.