feat(settings): update return types for `set` function of Setting class
Historically, the `set` functions in the Setting subclasses and other
functions in settings.js returned a boolean to indicate if the provided
value was valid and applied, or not.

They now return the value effectively set, which may differ from the
provided value if necessary (i.e. rounded or reset to default).